Complete wet part

+++ include/linux/syscalls.h
@@ -941,7 +941,7 @@ asmlinkage long sys_pkey_free(int pkey);
 asmlinkage long sys_statx(int dfd, const char __user *path, unsigned flags,
 			  unsigned mask, struct statx __user *buffer);
 asmlinkage long sys_hello(void);
-asmlinkage int  sys_set_status(int status);
+asmlinkage long  sys_set_status(int status);
 asmlinkage int  sys_get_status(void);
 asmlinkage int  sys_register_process(void);
 asmlinkage long sys_get_all_cs(void);
